Extract pattern overlap analysis into shared utility
